+++ contracts/ClearingHouse.sol
@@ -125,8 +125,8 @@ contract ClearingHouse is
         require(clearingHouseConfigArg.isContract(), "CH_CCNC");
         // AccountBalance is not contract
         require(accountBalanceArg.isContract(), "CH_ABNC");
-        // CH_ANC: address is not contract
-        require(exchangeArg.isContract(), "CH_ANC");
+        // CH_ENC: Exchange is not contract
+        require(exchangeArg.isContract(), "CH_ENC");
         // CH_IFANC: InsuranceFund address is not contract
         require(insuranceFundArg.isContract(), "CH_IFANC");
 
@@ -151,8 +151,8 @@ contract ClearingHouse is
 
     // solhint-disable-next-line func-order
     function setTrustedForwarder(address trustedForwarderArg) external onlyOwner {
-        // CH_ANC: address is not contract
-        require(trustedForwarderArg.isContract(), "CH_ANC");
+        // CH_TFNC: TrustedForwarder is not contract
+        require(trustedForwarderArg.isContract(), "CH_TFNC");
         _setTrustedForwarder(trustedForwarderArg);
         emit TrustedForwarderChanged(trustedForwarderArg);
     }
@@ -193,8 +193,8 @@ contract ClearingHouse is
                 })
             );
 
-        // price slippage check
-        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSC");
+        // CH_PSCF: price slippage check fails
+        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSCF");
 
         // if !useTakerBalance, takerBalance won't change, only need to collects fee to oweRealizedPnl
         if (params.useTakerBalance) {
@@ -331,8 +331,8 @@ contract ClearingHouse is
 
         int256 realizedPnl = _settleBalanceAndRealizePnl(trader, params.baseToken, response);
 
-        // price slippage check
-        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSC");
+        // CH_PSCF: price slippage check fails
+        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSCF");
 
         emit LiquidityChanged(
             trader,
@@ -491,8 +491,8 @@ contract ClearingHouse is
         //   trader: here
         //   baseToken: in Exchange.settleFunding()
 
-        // CH_NEO: not empty order
-        require(!IAccountBalance(_accountBalance).hasOrder(trader), "CH_NEO");
+        // CH_CLWTISO: cannot liquidate when there is still order
+        require(!IAccountBalance(_accountBalance).hasOrder(trader), "CH_CLWTISO");
 
         // CH_EAV: enough account value
         require(
@@ -846,8 +846,7 @@ contract ClearingHouse is
     }
 
     function _requireEnoughFreeCollateral(address trader) internal view {
-        // freeCollateral is calculated based on imRatio
-        // CH_NEFCI: not enough account value by imRatio
+        // CH_NEFCI: not enough free collateral by imRatio
         require(
             _getFreeCollateralByRatio(trader, IClearingHouseConfig(_clearingHouseConfig).getImRatio()) >= 0,
             "CH_NEFCI"