+++ contracts/Vault.sol
@@ -80,14 +80,14 @@ contract Vault is IVault, ReentrancyGuardUpgradeable, OwnerPausable, BaseRelayRe
     }
 
     function setTrustedForwarder(address trustedForwarderArg) external onlyOwner {
-        // V_ANC: TrustedForwarder address is not contract
-        require(trustedForwarderArg.isContract(), "V_ANC");
+        // V_TFNC: TrustedForwarder address is not contract
+        require(trustedForwarderArg.isContract(), "V_TFNC");
         _setTrustedForwarder(trustedForwarderArg);
     }
 
     function setClearingHouse(address clearingHouseArg) external onlyOwner {
-        // V_ANC: address is not contract
-        require(clearingHouseArg.isContract(), "V_ANC");
+        // V_CHNC: ClearingHouse is not contract
+        require(clearingHouseArg.isContract(), "V_CHNC");
         _clearingHouse = clearingHouseArg;
     }