+++ CHANGELOG.md
@@ -20,18 +20,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rename `AccountBalance.addOwedRealizedPnl` to `AccountBalance.modifyOwedRealizedPnl`
 - rename param `delta` of `AccountBalance.modifyOwedRealizedPnl` as `amount`
 - rename the param `sqrtPriceAfter` in the `ClearingHouse.PositionChanged` event to `sqrtPriceAfterX96`
+- rename error codes in `ClearingHouse`
+    1. `CH_NEO` to `CH_CLWTISO`
+    2. `CH_PSC` to `CH_PSCF`
+    3. `CH_ANC` to `CH_ENC`
+    4. `CH_ANC` to `CH_TFNC`
 - rename params of `ClearingHouse.openPosition`
     1. `deltaBase` to `base` 
     2. `deltaQuote` to `quote` 
 - rename params of `ClearingHouse.closePosition`
     1. `deltaBase` to `base` 
     2. `deltaQuote` to `quote` 
+- rename error code in `Exchange`: `EX_ANC` to `EX_BNC`
 - rename params in struct `Exchange.SwapResponse`
     1. `deltaAvailableBase` to `base` 
     2. `deltaAvailableQuote` to `quote` 
 - rename params in struct `Exchange.RealizePnlParams`
     1. `deltaAvailableBase` to `base` 
     2. `deltaAvailableQuote` to `quote` 
+- rename `OrderBook.getOwedFee` as `OrderBook.getPendingFee`
 - rename params in struct `OrderBook.RemoveLiquidityResponse`
     1. `deltaTakerBase` to `takerBase` 
     2. `deltaTakerQuote` to `takerQuote` 
@@ -41,6 +48,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rename params in struct `AccountMarket.Info`
     1. `takerBaseBalance` to `takerPositionSize` 
     2. `takerQuoteBalance` to `takerOpenNotional` 
+- rename error codes in `Vault` 
+    1. `V_ANC` to `V_CHNC`
+    2. `V_ANC` to `V_TFNC`
 
 - move event `FundingPaymentSettled` to ClearingHouse